requirements.txt
pymemcache
requests
discord.py
